requirements.txt
chromadb>=1.0.0
sentence-transformers>=2.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
